fix(onchange): harden endpoint + evaluator per PR #191 review
Six findings from the CodeRabbit review of PR #191 (onchange Phase 1):

1. Authorize before revealing entity/field existence
   (onchange-api.ts). Permission slot (via commandLayer.execute with
   skipActionSlots: true) now runs before the 400/404 existence
   checks, so unauthenticated probes can't enumerate which entities
   have onchange hooks — they get a uniform 403-shape envelope.

2. Reject malformed `values` (onchange-api.ts). Arrays / strings /
   nulls / numbers now yield 400 INVALID_REQUEST.MALFORMED_VALUES
   instead of silently coercing to `{}`. `undefined` still defaults
   to `{}` to preserve the "no initial values" case.

3. Drop late results from timed-out hooks (onchange-evaluator). Added
   `TimedHookOutcome { result, timedOut }` — when Promise.race
   resolves via timeout, the evaluator drops the late hook result,
   late warnings, and does not enqueue cascaded fields even if the
   background promise eventually resolves. Chose the `timedOut` flag
   approach over AbortController to keep OnchangeContext additive;
   no existing hook needs to cooperate with a signal.

4. Sanitize user-facing warnings (onchange-evaluator). Raw internal
   messages (SQL text, stack fragments) no longer leak into the
   response `warnings` array. Clients see shapes like `lookup on "X"
   failed` / `permission check failed for "X"`. Full error context is
   logged via the runtime Logger at warn level with entity/id/field/
   actor/tenant — debugging signal preserved server-side, no client
   exposure. Logger defaults to consoleLogger, overridable via
   OnchangeEvaluatorOptions.logger.

5. Defensively clone values across the evaluator/hook boundary
   (onchange-evaluator). Added `safeClone` using structuredClone with
   identity fallback. `ctx.values` per hook, `record[field]` from
   lookup, and `rows` from query are cloned so a misbehaving hook
   cannot mutate another hook's view or the shared accumulator. The
   DataProvider itself is not cloned (expensive + pointless at that
   depth — we only clone at the call-site boundary).

6. Split the 836-line onchange-evaluator test file. Now 6 files,
   each under 500 lines, grouped by concern: core, chaining/errors,
   permissions, validation, mutation, plus a shared fixtures module.
   No test case lost. Source file also split into
   `onchange-evaluator.ts` (public entry + docstring, 259 lines) and
   `onchange-evaluator-internals.ts` (helpers, 360 lines, explicitly
   marked as module-private) to stay under the 500-line limit after
   the fix-related growth.

Test count +17 (7 new evaluator cases for findings 3–5, 10 new
onchange-api cases for findings 1–2).

Closes #192.
