+++ addons/adapter-server/cap-adapter-server/__tests__/onchange-api.test.ts
@@ -190,7 +190,10 @@ describe("POST /api/entities/:name/onchange", () => {
     expect(body.success).toBe(false);
   });
 
-  test("404 when entity does not exist", async () => {
+  test("404 when entity does not exist (with permission allowed)", async () => {
+    // Permission slot is a no-op allow_all — with authorization in hand, the
+    // evaluator DOES reveal that the entity is missing. Only unauthenticated
+    // callers see a uniform 403 (see Finding 1 tests below).
     const { status, body } = await postOnchange(happyPort, "ghost", {
       changedField: "x",
       values: {},
@@ -241,6 +244,117 @@ describe("POST /api/entities/:name/onchange", () => {
     expect(body).toHaveProperty("warnings");
     expect(typeof body.updates).toBe("object");
     expect(Array.isArray(body.warnings)).toBe(true);
+  });
+});
+
+// ── Finding 1: auth before existence ──────────────────────
+
+describe("POST /api/entities/:name/onchange — authorize before revealing existence (Finding 1)", () => {
+  test("probe against non-existent entity with denying permission returns 403 (not 404)", async () => {
+    // The deny-all server rejects at the permission slot. The caller must
+    // NOT be able to tell from the response whether "ghost" exists — they
+    // only learn that they're not authorized.
+    const { status, body } = await postOnchange(denyPort, "ghost", {
+      changedField: "x",
+      values: {},
+    });
+    expect(status).toBe(403);
+    expect(body.success).toBe(false);
+    const err = body.error as Record<string, unknown>;
+    // Must not leak entity-existence information.
+    expect(JSON.stringify(err)).not.toContain("not found");
+    expect(JSON.stringify(err)).not.toContain("no onchange");
+  });
+
+  test("probe against real entity with denying permission returns 403", async () => {
+    const { status, body } = await postOnchange(denyPort, "purchase_line", {
+      changedField: "product_id",
+      values: { product_id: "p1" },
+    });
+    expect(status).toBe(403);
+    expect(body.success).toBe(false);
+  });
+
+  test("probe with missing changedField on denying permission still returns 403 (no 400 leak)", async () => {
+    // Previously the 400 branch for missing changedField ran before the
+    // permission slot, letting an unauthenticated caller distinguish "bad
+    // request shape" from "entity does not exist" vs "permission denied".
+    // Now auth runs first so the answer is always 403.
+    const { status } = await postOnchange(denyPort, "purchase_line", {
+      values: {},
+    });
+    expect(status).toBe(403);
+  });
+
+  test("probe with malformed values on denying permission still returns 403 (auth wins)", async () => {
+    const { status } = await postOnchange(denyPort, "purchase_line", {
+      changedField: "product_id",
+      values: "not an object",
+    });
+    expect(status).toBe(403);
+  });
+});
+
+// ── Finding 2: reject malformed `values` ──────────────────
+
+describe("POST /api/entities/:name/onchange — malformed values (Finding 2)", () => {
+  test("values as array → 400 with MALFORMED_VALUES code", async () => {
+    const { status, body } = await postOnchange(happyPort, "purchase_line", {
+      changedField: "product_id",
+      values: [],
+    });
+    expect(status).toBe(400);
+    const err = body.error as Record<string, unknown>;
+    expect(err.code).toBe("INVALID_REQUEST.MALFORMED_VALUES");
+  });
+
+  test("values as string → 400 with MALFORMED_VALUES code", async () => {
+    const { status, body } = await postOnchange(happyPort, "purchase_line", {
+      changedField: "product_id",
+      values: "not an object",
+    });
+    expect(status).toBe(400);
+    const err = body.error as Record<string, unknown>;
+    expect(err.code).toBe("INVALID_REQUEST.MALFORMED_VALUES");
+  });
+
+  test("values as null → 400 with MALFORMED_VALUES code", async () => {
+    const { status, body } = await postOnchange(happyPort, "purchase_line", {
+      changedField: "product_id",
+      values: null,
+    });
+    expect(status).toBe(400);
+    const err = body.error as Record<string, unknown>;
+    expect(err.code).toBe("INVALID_REQUEST.MALFORMED_VALUES");
+  });
+
+  test("values as number → 400 with MALFORMED_VALUES code", async () => {
+    const { status, body } = await postOnchange(happyPort, "purchase_line", {
+      changedField: "product_id",
+      values: 42,
+    });
+    expect(status).toBe(400);
+    const err = body.error as Record<string, unknown>;
+    expect(err.code).toBe("INVALID_REQUEST.MALFORMED_VALUES");
+  });
+
+  test("values as plain object → proceeds normally", async () => {
+    const { status, body } = await postOnchange(happyPort, "purchase_line", {
+      changedField: "product_id",
+      values: { product_id: "p1", quantity: 1 },
+    });
+    expect(status).toBe(200);
+    expect(body).toHaveProperty("updates");
+    expect(body).toHaveProperty("warnings");
+  });
+
+  test("values omitted entirely → defaults to {} and proceeds", async () => {
+    // `values` is optional — the old behavior of treating absent as `{}` is
+    // preserved. Only explicitly-wrong shapes trigger MALFORMED_VALUES.
+    const { status } = await postOnchange(happyPort, "purchase_line", {
+      changedField: "product_id",
+    });
+    expect(status).toBe(200);
   });
 });
